added profile wizards and updated with sub compoenents
