+++ frontend/src/containers/talent/profile/build-profile/contact-info/myContactInfo.css
@@ -9,7 +9,7 @@
 }
 
 
-@media only screen and (max-width: 1300px) {
+@media only screen and (max-width: 1680px) {
   .contact-info-view-container{
     padding-top: 30px;
     padding-left: 0px;
