+++ frontend/src/containers/talent/profile/build-profile/pictures/myPicturesScreen.css
@@ -1,12 +1,13 @@
 .profile-picture-image {
   width: 150px;
   height: 180px;
+  object-fit: cover;
 }
 
 .profile-picture-info-view-container {
   padding-top: 33px;
-  padding-left: 50px;
-  padding-right: 50px;
+  padding-left: 20px;
+  padding-right: 20px;
 }
 
 
@@ -18,7 +19,7 @@
   }
 }
 
-@media only screen and (max-width: 1400px) {
+@media only screen and (max-width: 1500px) {
   .profile-picture-info-view-container{
     padding-top: 33px;
     padding-left: 0px;
@@ -73,8 +74,8 @@
   background: red;
   font-weight: 600;
   position: relative;
-  top: 35px;
-  left: 59px;
+  top: 27px;
+  left: 60px;
 }
 
 .profile-picture-delete-icon:hover{
