+++ frontend/src/containers/talent/profile/build-profile/resume/myResumeScreen.css
@@ -1,6 +1,7 @@
 .profile-resume-image {
   width: 300px;
   height: 360px;
+  object-fit: cover;
 }
 
 .profile-resume-delete-icon {
